refactor(common): split god modules models.py and fields.py (#581)
Pure mechanical reorganization — no behavior change, no schema migration.

common/models.py -> common/models/ package, grouped by domain:
  base.py (TimeStampedModel, StripeConnectMixin, ExifStripMixin),
  site.py (Legal, SiteSettings), email.py (EmailLog),
  tags.py (Tag, TagAssignment, TagManager, TaggableMixin),
  files.py (FileExport, FileUploadAudit, QuarantinedFile),
  exchange.py (ExchangeRate). All re-exported from __init__ so every
  `from common.models import X` path (and the `common.models.SiteSettings`
  mock-patch targets) keeps working unchanged.

common/fields.py -> sanitizers extracted to common/sanitizers.py
  (sanitize_html, sanitize_markdown, render_markdown + allowlists).
  Field classes stay in fields.py, which re-exports the sanitizer API
  for backward-compatible imports.

makemigrations --check reports no new migrations; make check (ruff,
mypy --strict, file-length, i18n) and the common test suite pass.
