--- src/common/models.py
@@ -1,482 +1,0 @@
-import gzip
-import typing as t
-import uuid
-
-from django.conf import settings
-from django.contrib.contenttypes.fields import GenericForeignKey, GenericRelation
-from django.contrib.contenttypes.models import ContentType
-from django.core.validators import MaxValueValidator, MinValueValidator
-from django.db import models
-from simple_history.models import HistoricalRecords
-from solo.models import SingletonModel
-
-from .fields import MarkdownField, ProtectedFileField
-
-
-class TimeStampedModel(models.Model):
-    id = models.UUIDField(primary_key=True, default=uuid.uuid4, editable=False)
-    created_at = models.DateTimeField(auto_now_add=True, db_index=True)
-    updated_at = models.DateTimeField(auto_now=True, db_index=True)
-
-    class Meta:
-        abstract = True
-
-    def save(self, *args: t.Any, **kwargs: t.Any) -> None:
-        """Override the save method to call full_clean before saving."""
-        self.full_clean()
-        super().save(*args, **kwargs)
-
-
-class StripeConnectMixin(models.Model):
-    """Mixin for models that can connect a Stripe account (e.g., Organization, RevelUser).
-
-    Provides the four Stripe Connect fields and an ``is_stripe_connected`` property.
-    """
-
-    stripe_account_email = models.EmailField(null=True, blank=True, db_index=True)
-    stripe_account_id = models.CharField(
-        max_length=255,
-        blank=True,
-        null=True,
-        unique=True,
-        help_text="The Stripe Connect Account ID.",
-    )
-    stripe_charges_enabled = models.BooleanField(default=False)
-    stripe_details_submitted = models.BooleanField(default=False)
-
-    class Meta:
-        abstract = True
-
-    @property
-    def is_stripe_connected(self) -> bool:
-        """Check if the Stripe account is fully connected."""
-        return self.stripe_account_id is not None and self.stripe_charges_enabled and self.stripe_details_submitted
-
-    def _stripe_update_fields(self, *fields: str) -> list[str]:
-        """Build an ``update_fields`` list, appending ``updated_at`` when the model has it.
-
-        Models inheriting from ``TimeStampedModel`` have ``auto_now=True`` on
-        ``updated_at``, which is only honoured when it appears in ``update_fields``.
-        ``RevelUser`` (via ``AbstractUser``) does not have this field.
-        """
-        result = list(fields)
-        if hasattr(self, "updated_at"):
-            result.append("updated_at")
-        return result
-
-
-class ExifStripMixin(models.Model):
-    """Mixin that strips EXIF metadata from image fields on save.
-
-    Subclasses must define IMAGE_FIELDS as an iterable of field names to process.
-    """
-
-    IMAGE_FIELDS: t.Iterable[str]
-
-    def _strip_exif_from_image_fields(self) -> None:
-        from django.core.exceptions import ValidationError
-        from django.utils.translation import gettext_lazy as _
-        from PIL import UnidentifiedImageError
-
-        from common.utils import strip_exif
-
-        for field_name in self.IMAGE_FIELDS:
-            file = getattr(self, field_name, None)
-            if file:
-                try:
-                    setattr(self, field_name, strip_exif(file))
-                except (UnidentifiedImageError, OSError) as e:
-                    # Convert PIL errors to ValidationError for proper 400 response.
-                    # This is a safety net - validation should happen before save().
-                    raise ValidationError({field_name: [_("File is not a valid image.")]}) from e
-
-    def save(self, *args: t.Any, **kwargs: t.Any) -> None:
-        """Override save to auto-strip exif from image fields."""
-        self._strip_exif_from_image_fields()
-        super().save(*args, **kwargs)
-
-    class Meta:
-        abstract = True
-
-
-class Legal(SingletonModel):
-    """Singleton model for legal documents like Terms and Conditions and Privacy Policy."""
-
-    terms_and_conditions = MarkdownField(
-        help_text="Terms and Conditions text in markdown format", blank=True, default=""
-    )
-    privacy_policy = MarkdownField(help_text="Privacy Policy text in markdown format", blank=True, default="")
-    updated_at = models.DateTimeField(auto_now=True)
-    history = HistoricalRecords()
-
-    def __str__(self) -> str:  # pragma: no cover
-        return "Legal Documents"
-
-    class Meta:
-        verbose_name = "Legal Documents"
-
-
-class SiteSettings(SingletonModel):
-    """Singleton model for common application settings."""
-
-    class BannerSeverity(models.TextChoices):
-        DEBUG = "debug"
-        INFO = "info"
-        WARNING = "warning"
-        ERROR = "error"
-        CRITICAL = "critical"
-
-    notify_user_joined = models.BooleanField(
-        default=False, help_text="Send a notification when a new user joins the platform."
-    )
-    notify_organization_created = models.BooleanField(
-        default=False, help_text="Send a notification when a new organization is created."
-    )
-
-    created_at = models.DateTimeField(auto_now_add=True)
-    updated_at = models.DateTimeField(auto_now=True)
-    data_retention_days = models.PositiveIntegerField(
-        verbose_name="Data Retention Period",
-        help_text="The number of days to retain data before deletion.",
-        default=30,
-    )
-    live_emails = models.BooleanField(default=False, help_text="Live-emails enabled")
-    frontend_base_url = models.URLField(default=settings.FRONTEND_BASE_URL)
-    internal_catchall_email = models.EmailField(
-        verbose_name="Internal Catchall Email",
-        help_text="The catchall email address for internal use.",
-        default=settings.INTERNAL_CATCHALL_EMAIL,
-    )
-
-    # Platform business details for VAT invoicing
-    platform_business_name = models.CharField(
-        max_length=255,
-        blank=True,
-        default="",
-        help_text="Legal business name for invoices (e.g., 'Revel S.r.l.').",
-    )
-    platform_business_address = models.TextField(
-        blank=True,
-        default="",
-        help_text="Full business address for invoices.",
-    )
-    platform_vat_country = models.CharField(
-        max_length=2,
-        blank=True,
-        default="",
-        help_text="ISO 3166-1 alpha-2 country code for platform's VAT registration.",
-    )
-    platform_vat_id = models.CharField(
-        max_length=20,
-        blank=True,
-        default="",
-        help_text="Platform's VAT ID including country prefix (e.g., IT12345678901).",
-    )
-    platform_vat_rate = models.DecimalField(
-        max_digits=5,
-        decimal_places=2,
-        default=0,
-        validators=[MinValueValidator(0), MaxValueValidator(100)],
-        help_text="Platform's domestic VAT rate (e.g., 22.00 for 22%).",
-    )
-
-    # Invoice settings
-    platform_invoice_bcc_email = models.EmailField(
-        blank=True,
-        default="",
-        help_text="BCC email for all platform fee invoices (internal accounting copy).",
-    )
-
-    # Maintenance banner
-    maintenance_message = models.CharField(
-        max_length=500,
-        blank=True,
-        default="",
-        help_text="Message to display in the maintenance banner. Leave blank to disable.",
-    )
-    maintenance_severity = models.CharField(
-        max_length=20,
-        choices=BannerSeverity.choices,
-        default=BannerSeverity.INFO,
-        help_text="Severity level of the maintenance banner.",
-    )
-    maintenance_scheduled_at = models.DateTimeField(
-        null=True,
-        blank=True,
-        help_text="When the maintenance window starts.",
-    )
-    maintenance_ends_at = models.DateTimeField(
-        null=True,
-        blank=True,
-        help_text="When the maintenance window ends. Banner is hidden after this time.",
-    )
-
-    history = HistoricalRecords()
-
-    def clean(self) -> None:
-        """Normalize maintenance_message so whitespace-only values become empty."""
-        if self.maintenance_message:
-            self.maintenance_message = self.maintenance_message.strip()
-
-    @property
-    def is_maintenance_banner_active(self) -> bool:
-        """Whether the maintenance banner should currently be displayed."""
-        from django.utils import timezone
-
-        if not self.maintenance_message or not self.maintenance_message.strip():
-            return False
-        if self.maintenance_ends_at and self.maintenance_ends_at < timezone.now():
-            return False
-        return True
-
-    def __str__(self) -> str:  # pragma: no cover
-        return "Common Settings"
-
-    class Meta:
-        verbose_name = "Common Settings"
-        verbose_name_plural = "Common Settings"
-
-
-class EmailLog(TimeStampedModel):
-    to = models.EmailField(db_index=True)
-    subject = models.TextField(db_index=True)
-    sent_at = models.DateTimeField(auto_now_add=True, db_index=True)
-    test_only = models.BooleanField(default=False, db_index=True)
-    compressed_body = models.BinaryField(null=True, blank=True)
-    compressed_html = models.BinaryField(null=True, blank=True)
-
-    def set_body(self, body: str) -> None:
-        """Compress and set text."""
-        self.compressed_body = gzip.compress(body.encode())
-
-    def set_html(self, html_body: str) -> None:
-        """Compress and set html."""
-        self.compressed_html = gzip.compress(html_body.encode())
-
-    @property
-    def body(self) -> str | None:
-        """Decompress and return text."""
-        if self.compressed_body:
-            return gzip.decompress(self.compressed_body).decode()
-        return None
-
-    @property
-    def html(self) -> str | None:
-        """Decompress and return html."""
-        if self.compressed_html:
-            return gzip.decompress(self.compressed_html).decode()
-        return None
-
-    def __str__(self) -> str:
-        return f"Email to: {self.to}"
-
-    class Meta:
-        indexes = [
-            models.Index(fields=["to", "sent_at", "test_only"], name="ix_emaillog_to_sentat_testonly"),
-        ]
-
-
-# ---- Tag Models ----
-
-
-class Tag(TimeStampedModel):
-    name = models.CharField(max_length=64, unique=True, db_index=True)
-    description = models.TextField(blank=True, null=True)
-    color = models.CharField(max_length=7, blank=True, null=True, help_text="Hex color (e.g. #FF0099)")
-    icon = models.CharField(max_length=64, blank=True, null=True, help_text="Optional icon name")
-    parent = models.ForeignKey("self", null=True, blank=True, on_delete=models.SET_NULL, related_name="children")
-
-    @staticmethod
-    def normalize_name(name: str) -> str:
-        """Strip surrounding whitespace and leading ``#`` (hashtag) characters."""
-        return name.strip().lstrip("#").strip()
-
-    def clean(self) -> None:
-        """Normalize the name before validation."""
-        if self.name:
-            self.name = self.normalize_name(self.name)
-
-    def __str__(self) -> str:
-        return self.name
-
-
-class TagAssignment(TimeStampedModel):
-    tag = models.ForeignKey(Tag, on_delete=models.CASCADE, related_name="assignments")
-    content_type = models.ForeignKey(ContentType, on_delete=models.CASCADE)
-    object_id = models.UUIDField()
-    content_object = GenericForeignKey("content_type", "object_id")
-    assigned_by = models.UUIDField(null=True, blank=True)  # Could link to user, if wanted
-
-    class Meta:
-        unique_together = ("tag", "content_type", "object_id")
-        indexes = [
-            models.Index(fields=["content_type", "object_id"]),
-        ]
-
-    def __str__(self) -> str:
-        return f"{self.tag} -> {self.content_object}"
-
-
-# ---- Taggable Mixin ----
-
-
-class TagManager:
-    def __init__(self, instance: models.Model):
-        """Init."""
-        self.instance = instance
-
-    def all(self) -> t.List[Tag]:
-        """Return all tags."""
-        return [ta.tag for ta in self.instance.tags.all()]  # type: ignore[attr-defined]
-
-    def add(self, *names: str) -> None:
-        """Add the tags."""
-        from django.contrib.contenttypes.models import ContentType  # local import to avoid circularity
-        from django.db.models import Q
-
-        from common.utils import get_or_create_with_race_protection
-
-        ct = ContentType.objects.get_for_model(self.instance.__class__)
-        for name in names:
-            name = Tag.normalize_name(name)
-            if not name:
-                continue  # skip blanks
-            tag, _ = get_or_create_with_race_protection(Tag, Q(name=name), {"name": name})
-            TagAssignment.objects.get_or_create(
-                tag=tag,
-                content_type=ct,
-                object_id=self.instance.pk,
-            )
-
-    def remove(self, *names: str) -> None:
-        """Remove the tags."""
-        from django.contrib.contenttypes.models import ContentType
-
-        ct = ContentType.objects.get_for_model(self.instance.__class__)
-        TagAssignment.objects.filter(
-            tag__name__in=names,
-            content_type=ct,
-            object_id=self.instance.pk,
-        ).delete()
-
-    def clear(self) -> None:
-        """Delete all tags."""
-        from django.contrib.contenttypes.models import ContentType
-
-        ct = ContentType.objects.get_for_model(self.instance.__class__)
-        TagAssignment.objects.filter(
-            content_type=ct,
-            object_id=self.instance.pk,
-        ).delete()
-
-
-class TaggableMixin(models.Model):
-    tags = GenericRelation(
-        TagAssignment,
-        related_query_name="%(class)s",
-        content_type_field="content_type",
-        object_id_field="object_id",
-    )
-
-    class Meta:
-        abstract = True
-
-    @property
-    def tags_manager(self) -> TagManager:
-        """Helper to get the tag manager."""
-        return TagManager(self)
-
-    def add_tags(self, *names: str) -> None:
-        """Add tags."""
-        self.tags_manager.add(*names)
-
-    def remove_tags(self, *names: str) -> None:
-        """Remove tags."""
-        self.tags_manager.remove(*names)
-
-    def clear_tags(self) -> None:
-        """Clear tags."""
-        self.tags_manager.clear()
-
-    def get_tags(self) -> t.List[Tag]:
-        """Get tags."""
-        return self.tags_manager.all()
-
-
-class FileExport(TimeStampedModel):
-    """Tracks async file export jobs (e.g. Excel exports)."""
-
-    class ExportStatus(models.TextChoices):
-        PENDING = "PENDING"
-        PROCESSING = "PROCESSING"
-        READY = "READY"
-        FAILED = "FAILED"
-
-    class ExportType(models.TextChoices):
-        QUESTIONNAIRE_SUBMISSIONS = "questionnaire_submissions"
-        ATTENDEE_LIST = "attendee_list"
-        REVENUE_VAT_REPORT = "revenue_vat_report"
-
-    requested_by = models.ForeignKey(settings.AUTH_USER_MODEL, on_delete=models.CASCADE, related_name="file_exports")
-    export_type = models.CharField(max_length=40, choices=ExportType.choices, db_index=True)
-    status = models.CharField(max_length=20, choices=ExportStatus.choices, default=ExportStatus.PENDING, db_index=True)
-    file = ProtectedFileField(upload_to="exports/", null=True, blank=True)
-    error_message = models.TextField(null=True, blank=True)
-    completed_at = models.DateTimeField(null=True, blank=True)
-    parameters = models.JSONField(default=dict, blank=True)
-
-    def __str__(self) -> str:
-        return f"FileExport({self.export_type}, {self.status})"
-
-
-class FileUploadAudit(TimeStampedModel):
-    class FileUploadAuditStatus(models.TextChoices):
-        PENDING = "PENDING"
-        CLEAN = "CLEAN"
-        MALICIOUS = "MALICIOUS"
-
-    app = models.CharField(max_length=64, db_index=True)
-    model = models.CharField(max_length=64, db_index=True)
-    instance_pk = models.UUIDField(db_index=True)
-    field = models.CharField(max_length=64, db_index=True)
-    file_hash = models.CharField(max_length=64, db_index=True)
-    uploader = models.EmailField(db_index=True)
-    status = models.CharField(
-        choices=FileUploadAuditStatus.choices, max_length=20, db_index=True, default=FileUploadAuditStatus.PENDING
-    )
-    notified = models.BooleanField(default=False)
-
-    def __str__(self) -> str:
-        return f"{self.app}.{self.model}.{self.field} ({self.status})"
-
-
-class QuarantinedFile(TimeStampedModel):
-    audit = models.OneToOneField(FileUploadAudit, on_delete=models.CASCADE)
-    # ProtectedFileField stores under protected/ so the file is only reachable via a
-    # signed, authorized URL (Caddy forward_auth) — never from the public /media/* path.
-    # Quarantined files hold the original (malicious / possibly PII-bearing) bytes and must
-    # not be world-readable like the previous plain FileField allowed.
-    file = ProtectedFileField(upload_to="quarantined_files")
-    findings = models.JSONField()
-
-    def __str__(self) -> str:
-        return f"Quarantined: {self.audit_id}"
-
-
-class ExchangeRate(TimeStampedModel):
-    """Daily exchange rates from frankfurter.app (ECB data).
-
-    Stores rates in the frankfurter response format: base currency with a JSON
-    object mapping target currencies to their rates.
-    """
-
-    base = models.CharField(max_length=3, default=settings.DEFAULT_CURRENCY)
-    date = models.DateField(db_index=True)
-    rates = models.JSONField(help_text="Mapping of currency code → rate relative to base")
-
-    class Meta:
-        constraints = [models.UniqueConstraint(fields=["base", "date"], name="unique_exchange_rate_per_day")]
-        ordering = ["-date"]
-
-    def __str__(self) -> str:
-        return f"{self.base} rates for {self.date} ({len(self.rates)} currencies)"
