+++ src/common/models/__init__.py
@@ -1,0 +1,29 @@
+"""Common models package.
+
+Models are grouped into focused modules by domain and re-exported here so that
+existing ``from common.models import X`` import paths keep working unchanged.
+"""
+
+from common.models.base import ExifStripMixin, StripeConnectMixin, TimeStampedModel
+from common.models.email import EmailLog
+from common.models.exchange import ExchangeRate
+from common.models.files import FileExport, FileUploadAudit, QuarantinedFile
+from common.models.site import Legal, SiteSettings
+from common.models.tags import Tag, TagAssignment, TaggableMixin, TagManager
+
+__all__ = [
+    "EmailLog",
+    "ExchangeRate",
+    "ExifStripMixin",
+    "FileExport",
+    "FileUploadAudit",
+    "Legal",
+    "QuarantinedFile",
+    "SiteSettings",
+    "StripeConnectMixin",
+    "Tag",
+    "TagAssignment",
+    "TagManager",
+    "TaggableMixin",
+    "TimeStampedModel",
+]
